* fix wait timeouts   * fix XImage format filtering   * fix several error handling corner cases   * fix c++ error handling   * add missing Window.h   * add better hierarchy to library sources   * build configuration fixes   * portability fixes   * packaging updates

+++ test/dbg_scan.cc
@@ -61,7 +61,7 @@ void scan_image (const char *filename)
     Magick::Image image;
     image.read(filename);
     string file = image.baseFilename();
-    unsigned baseidx = file.rfind('/');
+    size_t baseidx = file.rfind('/');
     if(baseidx != string::npos)
         file = file.substr(baseidx + 1, file.length() - baseidx - 1);
     ofstream svg((file + ".svg").c_str());
